Fix unit test, update change log, refactored documentation.
