+++ components/collector/src/source_collectors/file_source_collectors/generic_json.py
